Add files via upload
This code still needs finessing... when we merged files there were things missing and functionality which disappeared... things to note: the hover functionality (with the headers turning lime-green #32CD32) is missing. The submit button does not turn lime-green #32CD32 on both pages (I updated one I haven't had time to update the other)... the logo is missing as I turned it into an SVG file (from a .png file) to try to make it hi-res - but it is not working/visible. When I have made my changes I will push them there, but maybe start a new repository?